SegClip/requirements.txt
numpy~=1.24.2
scikit-learn~=1.2.2
transformers~=4.27.4
tqdm~=4.64.1
pandas~=1.5.3
pillow~=10.1
kornia~=0.6.12
scikit-image==0.25.2